store password and decrypt private key when creating secret
